restrictions

+++ app/datatables/articles_datatable.rb
@@ -20,9 +20,13 @@ class ArticlesDatatable
         [
           article.title,
           article.description,
-          link_to('Edit', @view.edit_article_path(article), class: "btn btn-xs btn-primary")+" | "+
-          link_to('Show', article, class: "btn btn-xs btn-info")+" | "+
-          link_to('Delete', article, method: :delete, data: {confirm: "Are you sure? "}, class: "btn btn-xs btn-danger")  
+          if @view.logged_in? && @view.current_user == article.user
+            link_to('Edit', @view.edit_article_path(article), class: "btn btn-xs btn-primary")+" | "+
+            link_to('Show', article, class: "btn btn-xs btn-info")+" | "+
+            link_to('Delete', article, method: :delete, data: {confirm: "Are you sure? "}, class: "btn btn-xs btn-danger")
+          else
+            link_to('Show', article, class: "btn btn-xs btn-info")
+          end
         ]
       end
     end
